Fix overlap of the footer with search results
The overlap only happens when search results are dynamically updated.
When new results are loaded the "push div" disappears and the footer is
no longer properly pushed down to under the results.

This approach switches from using a "push div" to specifying bottom
padding on the body. This works but has the same caveat as the
"push div" in that it requires a fixed size, in this case `4em` to mimic
the behavior with the "push div".

I want to thank @entuland (entuland on #css@freenode) for their advice.

Fixes #256